Update roi_param_search_selectSession.py
 Modified to re-do roi parameter search. Useful when conditions for auto_pre_cur has changed.
    Instead of treating having 9 thxxx folders as done,
    if thxxx is done for certain threshold then just reuse the results.
    Modify roi_detection function
    2021/10/06 JK

+++ suite2p_process_jk/roi_param_search_selectSession.py
@@ -10,6 +10,13 @@ Then, around the peak value y, try y-0.05:y+0.05:0.05.
 With 9 iterations, get 0.05 resolution of threshold
 
 2021/09/24 JK
+
+Updates:
+    Modified to re-do roi parameter search. Useful when conditions for auto_pre_cur has changed.
+    Instead of treating having 9 thxxx folders as done, 
+    if thxxx is done for certain threshold then just reuse the results.
+    Modify roi_detection function
+    2021/10/06 JK
 """
 removingSessions = {'025Upper': ['014', '016', '017','018','024','025','5555_001','5555_004','5555_014','5555_103','9999_1', '9999_2'],
                     '025Lower': ['011', '012', '016','025','5554_001','5554_003','5554_012','5554_013','5554_103','9998_1', '9998_2'],
@@ -53,37 +60,35 @@ def roi_detection(ops, db):
     opsPath = os.path.join(db['save_path0'], db['save_folder'])
     threshold = int(np.round(db['threshold_scaling']*100))
     subDir = f'{opsPath}/plane0/th{threshold:03}/'
-    opsfn = f'{subDir}ops.npy'
-    iscellfn = f'{subDir}iscell.npy'
-    if not os.path.isdir(subDir) or not os.path.isfile(opsfn) or not os.path.isfile(iscellfn):
-       # raise (f'Path already exists for threshold {threshold}')
-    # else:
-        if not os.path.isdir(subDir):
-           os.mkdir(subDir)
+    
+    if not os.path.isdir(subDir):
+       os.mkdir(subDir)
+    fnList = os.listdir(subDir)
+    fnexists = []
+    fnresults = ['F.npy', 'Fneu.npy', 'iscell.npy', 'spks.npy', 'stat.npy', 'ops.npy'] # to test done folders
+    for fn in fnresults:
+        if fn in fnList:
+            fnexists.append(1)
+        else:
+            fnexists.append(0)
+    if not all(fnexists):
         run_s2p(ops,db)
         flist = [f.name for f in os.scandir(f'{opsPath}/plane0') if f.is_file()]
         for fn in flist:
            if fn[-4:] != '.bin':
                shutil.copy(f'{opsPath}/plane0/{fn}', f'{subDir}{fn}')
+    else:
+        print(f'Threshold {threshold} was run for suite2p already. Skipping...')
     numCell, numNotCell = auto_pre_cur(subDir)
     return numCell, numNotCell
 
 #%% Basic settings
-h5Dir = 'E:/'
+h5Dir = 'D:/TPM/JK/h5/'
 fastDir = 'C:/JK/' # This better be in SSD
 
 mice = [25,  27,  30,  36,  37,  38,  39,  41,  52,  53,  54,  56]
 zoom = [2,   2,   2,   1.7, 1.7, 1.7, 1.7, 1.7, 1.7, 1.7, 1.7, 1.7]
 freq = [7.7, 7.7, 7.7, 7.7, 6.1, 6.1, 6.1, 6.1, 7.7, 7.7, 7.7, 7.7]
-
-ops = default_ops()
-ops['tau'] = 1.5
-ops['look_one_level_down'] = False
-ops['save_mat'] = False
-ops['save_NWB'] = False # for now. Need to set up parameters and confirm it works
-
-thFolderLen = 9 # to test done folders
-fnresults = ['F.npy', 'Fneu.npy', 'iscell.npy', 'spks.npy', 'stat.npy', 'data.bin', 'ops.npy'] # to test done folders
 
 ops = default_ops()
 ops['tau'] = 1.5
@@ -95,13 +100,15 @@ ops['two_step_registration'] = True
 ops['keep_movie_raw'] = True
 ops['smooth_sigma_time'] = 2
 ops['move_bin'] = True
+ops['save_mat'] = False
+ops['save_NWB'] = False # for now. Need to set up parameters and confirm it works
 
 ops['nbinned']: 10000  # max number of binned frames for cell detection        
 ops['max_overlap'] = 0.75  # cells with more overlap than this get removed during triage, before refinement
 ops['allow_overlap'] = False
 #%%
 # for mi in [0,3,8]:
-for mi in [8]:
+for mi in [0]:
     mouse = mice[mi]
     ops['fs'] = freq[mi]
     ops['zoom'] = zoom[mi]
@@ -164,54 +171,51 @@ for mi in [8]:
                 sn = sntemp.split('_')[1]
             
             if sn not in removingSessions[volumeName]:
-                # Check if ROI param search was done in this folder
-                # (# of thxxx folder == 9) & (isfile, F.npy, Fnew.npy, iscell.npy, spks.npy, stat.npy)
                 tempDir = os.path.join(planeDir, f'{sn}/plane0/')
                 ls = os.listdir(tempDir)
-                tempThLen = sum([os.path.isdir(os.path.join(tempDir, lstr)) for lstr in ls])
-                tempResultLen = len([b for b in fnresults if b in ls])
-                if not ((tempThLen == thFolderLen) & (tempResultLen == len(fnresults))): # this folder is not done yet
-                    # Check for registration. If not, run registration
-                    if not (os.path.isfile(f'{tempDir}ops.npy') & os.path.isfile(f'{tempDir}data.bin')):
-                        db = {'h5py': sftemp,
-                        'h5py_key': ['data'],
-                        'data_path': [],
+                # Check for registration. If not, run registration
+                if not (os.path.isfile(f'{tempDir}ops.npy') & os.path.isfile(f'{tempDir}data.bin')):
+                    db = {'h5py': sftemp,
+                    'h5py_key': ['data'],
+                    'data_path': [],
+                    'save_path0': planeDir,
+                    'save_folder': sn,
+                    'fast_disk': fastDir,
+                    'roidetect': False,
+                    }
+                    run_s2p(ops,db)
+                    rawbinFn = f'{planeDir}{sn}/plane0/data_raw.bin'
+                    os.remove(rawbinFn)
+
+                # Detect ROI's - 1st round. 0.2 resolution
+                numroi = {} # dict, paired between threshold and numCell, numNotCell
+                thresholdList = [int(th*10)/10 for th in np.linspace(0,0.8,5)]
+                for threshold in thresholdList:
+                    db = {'data_path': [],
+                        'do_registration': 0, # Forcing to not run registration
                         'save_path0': planeDir,
-                        'save_folder': f'{sn}',
-                        'fast_disk': f'{fastDir}',
-                        'roidetect': False,
-                        }
-                        run_s2p(ops,db)
-                        rawbinFn = f'{planeDir}{sn}/plane0/data_raw.bin'
-                        os.remove(rawbinFn)
-    
-                    # Detect ROI's - 1st round. 0.2 resolution
-                    numroi = {} # dict, paired between threshold and numCell, numNotCell
-                    thresholdList = [int(th*10)/10 for th in np.linspace(0,0.8,5)]
-                    for threshold in thresholdList:
-                        db = {'data_path': [],
-                            'do_registration': 0, # Forcing to not run registration
-                            'save_path0': planeDir,
-                            'save_folder': f'{sn}',
-                            'rerun_jk': 1, # Only for JK modification for suite2p v0.09
-                            'threshold_scaling': threshold
-                        }
-                        numCell, numNotCell = roi_detection(ops, db)
-                        numroi[threshold] = [numCell, numNotCell] 
-                    
-                    # Pick the best, and then run the 2nd round. 0.1 resolution
-                    numroiArray = np.array(list(numroi.values()))[:,0]
-                    maxind = np.argmax(numroiArray)
-                    maxthreshold = int(list(numroi.keys())[maxind]*10)/10
+                        'save_folder': sn,
+                        'rerun_jk': 1, # Only for JK modification for suite2p v0.09
+                        'threshold_scaling': threshold,
+                        'move_bin': False # to prevent s2p from trying moving data_raw.bin
+                    }
+                    numCell, numNotCell = roi_detection(ops, db)
+                    numroi[threshold] = [numCell, numNotCell] 
+                
+                # Pick the best, and then run the 2nd round. 0.1 resolution
+                numroiArray = np.array(list(numroi.values()))[:,0]
+                maxind = np.argmax(numroiArray)
+                maxthreshold = int(list(numroi.keys())[maxind]*10)/10
                 
                 thresholdList = [maxthreshold-0.1, maxthreshold+0.1]
                 for threshold in thresholdList:
                     db = {'data_path': [],
                         'do_registration': 0, # Forcing to not run registration
                         'save_path0': planeDir,
-                        'save_folder': f'{sn}',
+                        'save_folder': sn,
                         'rerun_jk': 1, # Only for JK modification for suite2p v0.09
-                        'threshold_scaling': threshold
+                        'threshold_scaling': threshold,
+                        'move_bin': False # to prevent s2p from trying moving data_raw.bin
                     }
                     numCell, numNotCell = roi_detection(ops, db)
                     numroi[threshold] = [numCell, numNotCell]
@@ -226,9 +230,10 @@ for mi in [8]:
                     db = {'data_path': [],
                         'do_registration': 0, # Forcing to not run registration
                         'save_path0': planeDir,
-                        'save_folder': f'{sn}',
+                        'save_folder': sn,
                         'rerun_jk': 1, # Only for JK modification for suite2p v0.09
-                        'threshold_scaling': threshold
+                        'threshold_scaling': threshold,
+                        'move_bin': False # to prevent s2p from trying moving data_raw.bin
                     }
                     numCell, numNotCell = roi_detection(ops, db)
                     numroi[threshold] = [numCell, numNotCell]
